final_project/django_requirements.txt
Django~=4.1.7
redis~=4.5.3
requests~=2.28.2
djangorestframework~=3.14.0
celery~=5.3.1
django-celery-results~=2.5.1
django-filter~=22.1
django-cors-headers~=3.14.0
djangorestframework-simplejwt~=5.2.2
djoser~=2.2.0
social-auth-app-django~=5.2.0
django-ckeditor~=6.5.1
django-filter~=22.1
Pillow~=9.3.0
